made only restructuring change to the code, no new code added, created new components and module folder
