+++ src/App.tsx
@@ -1,11 +1,10 @@
 import { useState } from "react";
+import { Patient, PatientsService } from "./modules/patients/patients";
+import { createNewPatientsApi } from "./modules/patients/patients_api";
+import PatientsLoader from "./modules/patients/patients_loader_button";
+import PatientsSearch from "./modules/patients/patients_search";
+import { Header, Footer, ContentBox, List } from "./components";
 import "./App.css";
-import { Patient, PatientsService } from "./patients/patients";
-import { createNewPatientsApi } from "./patients/patients_api";
-import PatientsLoader from "./patients/patients_loader_button";
-import PatientsSearch from "./patients/patients_search";
-import Header from "./header";
-import Footer from "./footer";
 
 const App = () => {
   const [patients, updatePatients] = useState<Patient[]>([]);
@@ -18,16 +17,18 @@ const App = () => {
       <div className="App-container">
         <Header />
         <div className="App-body">
-          <PatientsSearch
-            loadPatients={patientsApi.Search}
-            onResults={updatePatients}
-          />
+          <ContentBox>
+            <PatientsSearch
+              loadPatients={patientsApi.Search}
+              onResults={updatePatients}
+            />
+          </ContentBox>
           <h3>Or</h3>
           <PatientsLoader
             loadPatients={patientsApi.All}
             onLoaded={updatePatients}
           />
-          {patients.length > 0 && displayPatients(patients)}
+          {patients.length > 0 && <List patients={patients} />}
         </div>
         <Footer />
       </div>
@@ -36,15 +37,3 @@ const App = () => {
 };
 
 export default App;
-
-const displayPatients = (patients: Patient[]) => {
-  return (
-    <ul>
-      {patients.map((p, k) => (
-        <li style={{ listStyle: "none" }} key={k}>
-          ✅ {p.name}{" "}
-        </li>
-      ))}
-    </ul>
-  );
-};
