+++ src/components/index.tsx
@@ -1,0 +1,6 @@
+export { default as Header } from "./header";
+export { default as Footer } from "./footer";
+export { default as Loader } from "./loader";
+export { default as ToastBox } from "./toast";
+export { default as ContentBox } from "./contentbox";
+export { default as List } from "./list";
