+++ src/components/loader/index.tsx
