+++ src/modules/patients/patients_api.ts
@@ -1,4 +1,4 @@
-import { sleep } from "../utils/sleep";
+import { sleep } from "../../utils/sleep";
 import {
   isPatients,
   PatientsService,
